DGP-50
Add home page for trainers.
